network: allow configuring firewalld zone for virtual network bridge device
Since we're setting the zone anyway, it will be useful to allow
setting a different (custom) zone for each network. This will be done
by adding a "zone" attribute to the "bridge" element, e.g.:

   ...
   <bridge name='virbr0' zone='myzone'/>
   ...

If a zone is specified in the config and it can't be honored, this
will be an error. If no zone is specified and the default zone
("libvirt") can't be set, it will be ignored.

(BTW, Federico Simoncelli proposed a similar patch 5 1.2 years (!!)
ago, but I misunderstood its usefulness at first, and by the time I
did we were both too busy to revisit it. libvirt's code has changed so
much in the intervening time that it was simpler to just rewrite from
scratch)

Signed-off-by: Laine Stump <laine@laine.org>
Message-Id: <20190110025737.29755-6-laine@laine.org>