qemu_process: Expose process exit status code
Expose the process exit status code so the status code can be used to
determine if stderr (qmperr) should be logged.

In the case of virQEMUCapsInitQmp,
we only log stderr (qmperr) if the -first- query process fails with a
non zero status code.

Logging functions are pulled inside virQEMUCapsInitQmp where we can
detect that the -first- (not subsequent) QEMU process activation failed
with nonzero status.

Signed-off-by: Chris Venteicher <cventeic@redhat.com>
Message-Id: <20190113005032.17880-10-cventeic@redhat.com>