qemu_driver: Support baseline calculation using QEMU
Add capability to calculate hypervisor baseline using QMP message
exchanges with QEMU in addition to existing capability to calculate
baseline using libvirt utility functions.

A new utility function encapsulates the core logic for interacting with QEMU
using QMP baseline messages. The QMP messages only allow two cpu models
to be evaluated at a time so a sequence of QMP baseline messages must be
used to include all the models in the baseline when more than 2 cpu
models are input.

A QEMU process must be started prior to sending the series of QEMU
messages to compute baseline.

The QEMU process is started and maintained in the main hypervisor
baseline function and only a pointer to the QEMU process monitor is
passed to the baseline utility function.

The QEMU process is maintained in the main hypervisor baseline function
because the process will also be used for feature expansion (via QEMU)
in a later patch.

Signed-off-by: Chris Venteicher <cventeic@redhat.com>
Message-Id: <20190113005032.17880-32-cventeic@redhat.com>