docs, tests: Add schema, description, and tests for NFS namespace
Modify the storagepool.rng to allow for the usage of a different
XML namespace to parse the netfs_mount_opts to be included with
the netfs storage pool definition.

Modify the storagepoolxml2xmltest to utilize a properly modified
XML file to parse and format the namespace for a netfs storage pool.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20190116010922.28250-10-jferlan@redhat.com>