storage: Process storage pool capabilities
https://bugzilla.redhat.com/show_bug.cgi?id=1581670

During storage driver backend initialization, let's save
which backends are available in the storage pool capabilities.

In order to format those, we need add a connectGetCapabilities
processor to the storageHypervisorDriver. This allows a storage
connection, such as "storage:///system" to find the API and
format the results.

NB: This is not available from other hypervisor drivers such as
QEMU type connections.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20190116011549.16309-7-jferlan@redhat.com>