util: audit: Fix logging an error when kernel lacks audit support
Based on an upstream discussion, reporting the errno is useful for the
user to know why audit isn't supported. Even though having an error in
the logs might look concerning when 'audit_log=1', it also denotes that
audit is only going to be used if it's available, continuing normally
if it's unavailable for whatever reason.

Partially reverts commit 4199c2f221c.

https://bugzilla.redhat.com/show_bug.cgi?id=1596119

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <aa99f6c7600ca5fa098c1b97b0cf1ffaa3227852.1547650666.git.eskultet@redhat.com>