qemu: error out when vnc vncTLSx509secretUUID is unsupported
Add a capability check to qemuDomainDefValidate and refuse to start
a domain with VNC graphics if the TLS secret was set in qemu.conf
and it's not supported.

Note that qemuDomainSecretGraphicsPrepare does not generate any
secret data if the capability is not present and qemuBuildTLSx509BackendProps
is not called at all.

Signed-off-by: Ján Tomko <jtomko@redhat.com>
Message-Id: <447d472a6cd5bcd1d10a4580307da818ed3ffbbb.1548075388.git.jtomko@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>