util: fixing wrong assumption that PF has to have netdev assigned
libvirt wrongly assumes that VF netdev has to have the
netdev assigned to PF. There is no such requirement in SRIOV standard.
This patch change the virNetDevSwitchdevFeature() function to deal
with SRIOV devices which does not have netdev on PF. Also corrects
one comment about PF netdev assumption.

One example of such devices is ThunderX VNIC.
By applying this change, VF device is used for virNetlinkCommand() as
it is the only netdev assigned to VNIC.

Signed-off-by: Radoslaw Biernacki <radoslaw.biernacki@linaro.org>
[ dannf: Reinstated error path in virNetDevGetPhysicalFunction() as
  suggested by Michal Privoznik ]
Signed-off-by: dann frazier <dann.frazier@canonical.com>
Message-Id: <20190122192615.9256-2-dann.frazier@canonical.com>