qemu: domain: Allow overriding disk source in qemuDomainDetermineDiskChain
When we need to detect a chain for a image which will become the new
source for a disk (e.g. after a disk media change or a blockjob) we'd
need to replace disk->src temporarily to do so.

Move the 'disksrc' temporary variable as an argument and adjust callers.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <e1db4117f6aecce407f1197041c338d0e10b2ad7.1548259711.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>