security: Remove security driver internals for disk labelling
Security labelling of disks consists of labelling of the disk image
itself and it's backing chain. Modify
virSecurityManager[Set|Restore]ImageLabel to take a boolean flag that
will label the full chain rather than the top image itself.

This allows to delete/unify some parts of the code and will also
simplify callers in some cases.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <8d4d20a622ff06d82688ebf324f8940259bb02eb.1548259711.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>