qemu: Label backing chain of user-provided target of blockCopy when starting the job
Be more sensible when setting labels of the target of a
virDomainBlockCopy operation. Previously we'd relabel everything in case
it's a copy job even if there's no unlabelled backing chain. Since we
are also not sure whether the backing chain is shared we don't relabel
the chain on completion of the blockjob. This certainly won't play nice
with the image permission relabelling feature.

While this does not fix the case where the image is reused and has
backing chain it certainly sanitizes all the other cases. Later on it
will also allow to do the correct thing in cases where only one layer
was introduced.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <2ada12bb6966c8dd274978fdf85102d79b2d384e.1548259711.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>