Add newline after last entry in VIR_ENUM_IMPL calls
This is a follow on from a discussion with Andrea here:
https://www.redhat.com/archives/libvir-list/2019-January/msg00750.html

Most VIR_ENUM_IMPLS look like this:

VIR_ENUM_IMPL(XXX, XXX,
              "value1",
              "value2");

But he and I find this format nicer:

VIR_ENUM_IMPL(XXX, XXX,
              "value1",
              "value2",
);

The latter format means new enum additions only trigger 1 line of
diff, vs 2 for the current common method.

Since this patch series is already touching a lot of VIR_ENUM_IMPL
endings, now is as good a time as any to widely switch to that
format if others agree.

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Message-Id: <e781e13440cb0dcd32281340c06346ba164b294c.1548283535.git.crobinso@redhat.com>