virjson: always raise vir error on append failures
A function that returns -1 for multiple possible failures, but only
raises a libvirt error for some of those failures, is hard to use
correctly. Yet both of our JSON object/array appenders fall in that
pattern. We should either use the _QUIET memory allocation variants,
and make callers decide to report failure; or we can make all failure
paths noisy. This patch takes the latter approach.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190125040555.14382-2-eblake@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>