qemu: Assume migration with a network disk migration is safe
If a domain has a disk that is type='network' we require specific
cache mode to allow migration with it (either 'directsync' or
'none'). This doesn't make much sense since network disks are
supposed to be safe to migrate by default.

At the same time, we should be checking for the actual source
type, not apparent type set in the domain XML.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <f4d67fb62782b06c728bb8c9a10b4323cb11e516.1548752081.git.mprivozn@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>