qemu: blockjob: Don't report block job progress at 100% if job isn't ready
Some clients take the advice to poll virDomainGetBlockJobInfo rather
than wait for the ready event. In some cases qemu can get to 100% and
still not reach the synchronised phase.

Since we are dealing with a computer interacting, the error that the job
can't be finalized yet is not handled very well by those specific
implementations.

Our docs now correctly state to use the event.

We already munge the output for the start of the job, so we can do it
even here.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <cbaa0f0d6a2eafc9692bd349289d6c1356ab4f22.1548779049.git.pkrempa@redhat.com>